# Fix data frame initialisation in `bind_rows()`