Closes #5417.