@lionel- 2020-08-28
Performance impact:

```r
new_list <- purrr::map(1:10000, function(x) list(A = 2 * x, B = 4 * x)
bench::mark(bind_rows(new_list))[1:8]

#> # Master
#>   expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 bind_rows(new_list)    181ms    194ms      5.25     821KB     14.0     3     8

#> # This PR
#>   expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 bind_rows(new_list)    552ms    552ms      1.81     821KB     9.05     1     5
```