@lionel- 2021-04-20
@DavisVaughan Good point about name-repair. We don't need to repair names before calling `vec_rbind()` since `vec_rbind()` repairs them. Fixed in https://github.com/tidyverse/dplyr/commit/827615c591da2f77a24acc5dfcc0a2ec5d858ccd.

The perf example above was reduced from 370ms to 240ms. Thanks!